initial commit to run jobs via the ALCC allocation
1. edit setup.py
-- change version number to 2.0.0a (the idea is to keep this consistent with the version of number of the public release of OQMD itself).
-- add Python version (2.6--2.7) compatibility info.
-- edit install_requires to reflect changes in package names and certain packages which only play nice if the correct version is used.
2. edit calculation.py and structure.py to get the KPOINTS file from Tim Mueller's k-point server.
3. edit queue.py, resources.py, qfiles/edison_shared.q, and resources/*.yml to enable running jobs on Edison with our ALCC allocation.
4. (mostly Logan) edit scripts.py, vasp_settings/input/*.yml to enable HSE runs.
5. add safe switching in/out of folders (for running the Tim Mueller k-points script, which is distributed under
qmpy/analysis/vasp/getKPoints) with context handling.

+++ qmpy/analysis/vasp/calculation.py
@@ -428,6 +428,20 @@ class Calculation(models.Model):
                         Atom.objects.filter(id=atom.id).update(magmom=mom)
         self.settings = settings
 
+    def get_tm_kpoints(self):
+        try:
+            kpoints = self.input.get_tm_kpoint()
+            return kpoints
+        except:
+            kpts = self.input.get_kpoint_mesh(self.settings.get('kppra', 8000))
+            if self.settings.get('gamma', True):
+                kpoints = 'KPOINTS \n0 \nGamma\n'
+            else:
+                kpoints = 'KPOINTS \n0 \nMonkhorst-Pack\n'
+            kpoints += ' '.join( str(int(k)) for k in kpts ) + '\n'
+            kpoints += '0 0 0'
+            return kpoints
+
     def get_kpoints(self):
         kpts = self.input.get_kpoint_mesh(self.settings.get('kppra', 8000))
         if self.settings.get('gamma', True):
@@ -1416,29 +1430,32 @@ class Calculation(models.Model):
     def estimate(self):
         return 72*8*3600
 
+    _instruction = {}
     @property
     def instructions(self):
         if self.converged:
             return {}
+        
+        if not self._instruction:
+            self._instruction = {
+                    'path':self.path,
+                    'walltime':self.estimate,
+                    'header':'\n'.join(['gunzip -f CHGCAR.gz &> /dev/null',
+                        'date +%s',
+                        'ulimit -s unlimited']),
+                    'mpi':'mpirun -machinefile $PBS_NODEFILE -np $NPROCS',
+                    'binary':'vasp_53', 
+                    'pipes':' > stdout.txt 2> stderr.txt',
+                    'footer':'\n'.join(['gzip -f CHGCAR OUTCAR PROCAR ELFCAR',
+                        'rm -f WAVECAR CHG',
+                        'date +%s'])
+                    }
 
-        instruction = {
-                'path':self.path,
-                'walltime':self.estimate,
-                'header':'\n'.join(['gunzip -f CHGCAR.gz &> /dev/null',
-                    'date +%s',
-                    'ulimit -s unlimited']),
-                'mpi':'mpirun -machinefile $PBS_NODEFILE -np $NPROCS',
-                'binary':'vasp_53', 
-                'pipes':' > stdout.txt 2> stderr.txt',
-                'footer':'\n'.join(['gzip -f CHGCAR OUTCAR PROCAR ELFCAR',
-                    'rm -f WAVECAR CHG',
-                    'date +%s'])}
-
-        if self.input.natoms <= 4:
-            instruction.update({'mpi':'','binary':'vasp_53_serial',
-                'serial':True})
-        return instruction
-
+            if self.input.natoms <= 4:
+                self._instruction.update({'mpi':'','binary':'vasp_53_serial',
+                    'serial':True})
+        return self._instruction
+    
     def set_label(self, label):
         self.label = label
         if not self.entry is None:
